nixos: Removes zfs support from "new_kernel" installer variants
This is because of the phases where zfs support is not yet
available for the new kernel, forcing "new_kernel" variants not to be
built. This might be affecting aarch64-linux more often than
x86_64-linux.

The x86_64-linux image is seldom used, most likely uses are for
extremely new hardware exhibiting issues on the current LTS. The
aarch64-linux image is more likely to be used considering how board
support is always improving. It was the previous default on
aarch64-linux.