networkmanager: port to Meson
All hail Meson!

One serious issue is that building docs does not work.

We patch gobject-introspection to use absolute paths for shared libraries
in GIR files. Building the NetworkManager docs relies on the produced
introspection data but since the library is not yet installed
at the time the docs are generated, the build will fail.

It works in Autotools for some reason; they probably use
the pregenerated GIRs from the tarball.

Disabling the docs completely is not possible at the moment either,
since nmc [depends on them][1].

I have decided to fix this by pointing the installed location to the one
in the build directory using libredirect. Unfortunately, we cannot just set
the environment variables directly, since the build system runs
the documentation generator in a clean environment.

I have also added man, doc and devdoc outputs so the generated files have
somewhere to go.

Secondly, since Nix store is immutable, we also cannot use the package prefix
for configuration and mutable state data. At the same time, we cannot write
to the appropriate global directories during build. Autotools allowed to change
this in installFlags but Meson lacks similar mechanism so we need to patch
the build files.

Finally, I also removed the at_console patch since the permission has been
removed in 0.9.10.

[1]: https://bugzilla.gnome.org/show_bug.cgi?id=796755