pythonPackages.box2d: remove unnecessary pkgs-box2d buildInput
this dependency turns out to be completely unnecessary - the pypi
package includes its own bundled copy of the source (i know - ew.)
and building without pkgs-box2d results in a binary-identical .so.

this has the side-effect of enabling this package on darwin, which was
previously restricted by the apparently linux-only pkgs-box2d.