perl.meta.priority += 1
give priority to perl libraries when they meet the perl derivation in `buildEnv`.

The notable case is `buildEnv` inside `perl.withPackages`.

The `perl' derivation includes obsolete versions of some CPAN packages
which leads to collissions when there are newer versions
of the same libraries are on the right hand side
of `perl.withPackages` (perhaps indirectly).

Fixes #60025