docbook-xsl-nons: 1.79.1 → 1.79.2
Nothing much changed other than adding the suffix -nons to Docbook 4 styles,
making the Docbook 5 the variant without suffix:

https://github.com/docbook/xslt10-stylesheets/releases/tag/release%2F1.79.2

I have decided to make docbook_xsl an alias of docbook-xsl-nons for now,
to avoid confusion.

I also updated patch link that was broken to a newer version of the patch
and started sharing it between the two packages.