lorri: reserve attribute name
lorri is a nix-shell replacement for project development.
Since it’s public beta announcement was noticed by many people, they
are going to assume it is available from nixpkgs. We lead them to the
installation instructions while the tool is not yet ready for nixpkgs.

Related-issue: https://github.com/NixOS/nixpkgs/pull/60211