buildEnv: Don't warn about collisions if contents match when ignoreCollisions=true.
This prevents giving many false-positive collision warnings to NixOS users and makes it easier to find real collisions that should be addressed.

Contents are checked only if checkCollisionContents is true, which is the case for the NixOS system-path by default. If
checkCollisionContents is false, there is still a warning for every collision.