leksah: Point to upstream repository for using it
The build in nixpkgs was broken for forever. Trying to fix it would be
painful, as leksah uses a lot of overrides and custom versions. In
addition, leksah works best with the GHC version matching your project,
which means we'd have to make sure it builds for each of our GHC
versions.