pythonPackages.purepng: fix for py3.7, enable tests
the pypi package includes pre-cythonized versions of some components,
however these were cythonized with a py3.7-unaware cython which is what
caused the main breakage. fetching the equivalent version from github
allows us to offer either the pure-python version or one we cythonize
ourselves if the cython package is provided.

the added bonus of using a github checkout is that the tests are included,
so enable these.

these tests revealed that for full py3.7 compatibility an additional patch
is needed from an upstream PR.