liburing: bump, fix version (1.0.0pre92_7b989f3)
I accidentally got the number of commits wrong in the previous
prerelease version string. This is now fixed.

Generally, this would result in functions like builtins.compareVersions
to give incorrect results, so 'nix-env -u' doesn't work. But I'm
justifying it here, because: most people use it as a library, so the
hash change is all that matters. Plus, I only authored this a week or
so ago in upstream, so this change is fast enough that I think people
will be fine with it and can work around, especially since it's
unreleased in any stable channel.

This also bumps the library to the newest version, which contains some
bugfixes, and now installs the manpages into the $man output for us.

Signed-off-by: Austin Seipp <aseipp@pobox.com>