hplip: enable the plugin for aarch64
aarch64 is supported by the plugin since version 3.15.4 according to

  https://developers.hp.com/hp-linux-imaging-and-printing/release_notes

I have tested this change and it worked on my raspberry pi and with my
printer.

A minor caveat is that I had to revert to hplip 3.18.5 in order to be
able to build hplip itself for aarch64. The version 3.19.1 currently
fails to build on non-x86 because of

  https://bugs.launchpad.net/hplip/+bug/1785230

It is for this reason that I'm not adding aarch64-linux to the platforms
list for hplip in this commit.