memtest86: 4.3.6 -> 8.0
This updates memtest86 to (almost) the latest version.

memtest86-4.3.6 is the older version.  It was open source, but marked
broken.  memtest86-8.0 is the newer version.  It is no longer open
source, but it is possible to build and use.  There appear to be no
restrictions on its use or redistribution.