glibc: add patch to use utf-8 when LANG is unset
This allows the default builder to use utf-8 encoding. Previously,
when LANG and LC_ALL was left unset, glibc would default to “C”. This
is not good in many.

As a result of this, I think we can get rid of a lot of uses of LC_ALL
in Nixpkgs, where utf-8 was needed. These can be found by grepping
for:

  (?:LC_ALL|LANG) *= *["']?(?:en_US|C)\.UTF-8["']?