treewide: don’t assume we’re all americans
A couple of places set LANG=“en_US.UTF-8” to get unicode working in
the Nix builder. This is very very bad! Not everyone who uses Nix
lives in the U.S. and is an English speaker. Instead we should use the
“C.UTF-8”, which is an UTF-8 capable version of the unlocalized C
locale. Please use this one!

The previous commit makes Glibc fall back to this locale when LANG is
unset. This is in line with the Musl behavior. As a result, the Nix
builder is made UTF-8 capable without any LOCALE_ARCHIVE or LANG hacks.