scons: Fix a wrapping issue which overrides your PATH
This fixes the following situation:
If your PATH contains a python.withPackages interpreter and as part of
your SConstruct build you want to launch a python script in a
subprocess, which should run with '#!/usr/bin/env python', then it would
be problematic to have scons wrapped with an additional PATH pointing to
a different python interpreter. None of your withPackages-packages would
be importable.