xmind: 7.5-update1 -> 8-update8
The latest version of XMind doesn't support a prebuilt debian package
anymore, hence a manual installation of the ZIP is needed. The new
configuration for the Java GUI had to be patched as it relied on several
paths in the ZIP that won't work anymore in a Nix-based environment.

Additionally the following changes were made:

* Manual creation of a desktop item and an icon (we're using the one
  from AUR for now). Those files were available in the old `.deb`, but
  that isn't supported for the latest XMind version.

* Created a patch fro `XMind.ini` to use writable paths in $HOME and fix
  the path to the plugins in the store path.

* Created a custom startup script which ensures the creation of
  `$HOME/.xmind` and copies the relevant files from the store path into
  it. Such a behavior was available in the old `.deb` as well, but not
  anymore with the new ZIP.