nvidia-x11: Add patches for kernel 5.1.
Produced this on my own based on the errors and kernel changes:

https://github.com/torvalds/linux/commit/3d3539018d2cbd12e5af4a132636ee7fd8d43ef0
This is an intentional breaking change forcing drivers to acknowledge that the
return type of vm_vault handlers has changed to vm_fault_t and that VM_FAULT_*
must be returned. The driver only needs to be fixed to actually return vm_fault_t,
it already returns VM_FAULT_* values.

https://github.com/torvalds/linux/commit/fcd70cd36b9bf697122538c9e38e8cf954b2342b
Some DRM helper functions were moved to a separate header.