systemd: build with cryptsetup and cryptsetup-generators
We have to break the dependency chain mentioned in
b06cb3578d4ec2b76dcf2ad04486f3a1420c6c72.
(systemd -> cryptsetup -> lvm2 -> udev=systemd)

We do this by updating the lvm2 derivation to build without udev, so we
can give systemd a (lib)cryptsetup it can link against. We still need to
check whether this is enough, or if lvm needs to be bootstraped by a
udev/systemd which does not yet contain cryptsetup support - hopefully
not.

While working on the lvm2 derivation, the following cleanups where done
in the lvm2 derivation:
 - set installFlags directly, instead of doing it inside a script in
   preInstall
 - use installTargets again ("install", plus
   "install_systemd_{generators,units,configuration}" when udev is not
   null)
 - The call to the blkid binary in lvm2's 13-dm-disk.rules file
   disappeared (so we don't need to patch in blkid anymore).
   lvm seems to rely on udev's internal blkid functionality.
 - Call /run/current-system/systemd/bin/udevadm instead
   of ${systemd}/bin/udevadm in the lvm activation generator.
   This is not necessary to break the dependency cycle (as we don't use
   that file when building wihtout udev), but a good idea anyways -
   We want to trigger the udevadm of the current system, not the one
   that lvm was built with.
 - Remove the unused purity.patch

Co-Authored-By: Arian van Putten <aeroboy94@gmail.com>