cryptsetup-generator: remove
This package previously did override the systemd package, and instructed
ninja, systemd's previous build system, to only build the
cryptsetup-specific systemd generators (plus some rpath massaging, as we
didn't use ninja install).

Afterwards, users were expected to add this package to their
`systemd.generator-packages` NixOS module options, so systemd will use
these generators.

As the previous commit added cryptsetup support directly to the systemd
package (and pkgs.systemd now already ships the cryptsetup generators),
we don't need another package shipping the same generators.

So we remove the custom derivation. It might be possible to alias the
attribute to `systemd`, as adding systemd's generators to the list of
used generators should be a no-op, but it's cleaner to prompt users to
update their configuration.