clickshare-csc1: init at 01.07.00.033
Barco ClickShare is a wireless presentation system
where a USB dongle transmits to a base station
that is connected with a beamer.

Note that this package also contains a udev rule
file to grant access to the dongle if connected.
Access is limited to a group "clickshare"
(the name may be altered by package options).