Elm: fix "Map.! when compiling with --debug"
The elm compiler has a long standing bug that may prevent compiling in debugging mode.
The extremely slow pace of release means that this bug is open since months and will be open for many more.
While it seems like there will be a 0.19.1 release, we have no idea on the timeframe, so we apply a proposed patch in the meantime.
This patch is a fix for #1851.