fakechroot: 2.19 -> 2.20.1, prefer patch and official source over "fork"
The most immediate motivation for the source/patch reworking is to
upgrade the tree without needing to provide an updated version
with the fix and the upgrade as well.

Naturally the patch won't necessarily apply to all future versions,
but that's easy enough to check and happily appears to have no trouble
with this upgrade.