zoom-us: set LD_PRELOAD=.../v4l2convert.so
This was needed for my webcam to work with zoom-us. v4l2convert is a
wrapper for apps that only support simple bgr24 or yuv420. For these
apps to support webcams that don’t use those formats, we need to use
LD_PRELOAD to get those symbols. Taken from archlinux wiki:

https://wiki.archlinux.org/index.php/Webcam_setup#V4L1_support