mautrix-whatsapp: 2019-02-24 -> 2019-07-04
Bump to the latest revision of `mautrix-whatsapp` to regain
compatibility with matrix-synapse 0.99.5.

Please note that it was necessary to alter some of the sources in
`deps.nix`, please read the comment at the top of the file for further
information.

(cherry picked from commit b86a3e46b378a087ea57fd043fd59324665d24c9)