plan9port: build everything on Darwin
A commit in 2014 (see #2266) disabled building parts which depended on
X11 on Mac OS.  There's no explanation given in the pull request.  In
any case, many people install plan9port on Darwin specifically to use
the UI components.

Aside from this, the derivation was broken on Mac OS, probably because
people put the `which` and `perl` dependencies inside the !isDarwin
conditional.  Some programs worked, however the static libraries did
not build, so this could not be used as a buildInput to build programs
which need the libraries.  This fixes that.