nixos/nix: add support for specifying additional nix paths
It is a common pitfall for NixOS users to declare some additional nix
search paths by setting `nix.nixPath` not noticing that they are
removing the default search paths.

This change introduces the option `nix.extraNixPath` which can be used
to add paths to the list of search paths without overriding the
defaults. Users most likely want to use this since it is less likely
they'll have to resort fiddling the `NIX_PATH` environment variable
themselvs in order to repair their system.