docker-containers: split container creation / run stages
By doing this, we have systemd only supervise the actual run of the container, not the creation of it.
This may seem slightly silly, but if we ever wanted to add support for extra commands before / after container launch (like extra networks), this makes it super straightforward.
Plus, the idea of having the actual supervised start command be the one that _launches_ the container feels better?