nixos/network-interfaces: laziness fix
Apparently something else I'm working on made it a problem for config
declarations in this module to depend on evaluating utils.nix. I don't
understand where the evaluation cycle is, but in this case it's shorter
to reference `utils.` explicitly once than it is to say `with utils;`,
so I'm not bothering to think about it very hard.