buildGoPackage: keep string context (#63680)
In Nix, each string has a context that it carries of where it
originated. Some functions like filterAttrs modify the context of its
args when doing comparisons. That is important because we use the
string context of “name” to get where a derivation was defined. This
causes some builtins like unsafeGetAttrPos to report incorrectly that
the string was set in lib/attrsets.nix and reporting that as the
source file. Using removeAttrs avoids this problem.

Fixes #63679

(cherry picked from commit 1f46aaab1b679b4db0b78b3ee16c0978f0bf66aa)
The web list of packages shows the current stable branch, so we needed
to backport this to fix that effect soonish.  The change itself seems
very safe; only one rebuild is detected: `common-updater-scripts`